refactor: create new HeadProcessor class
